Sprinkle a lot of casts around to avoid generating random usize
That might help with portability, but it's quite a
hack. Realistically, these numbers should never get that big (we don't
really support mazes beyond hundreds of cells), but it's still
ugly. Maybe I should remove all uses of usize where I can…

+++ src/maze/algorithms.rs
@@ -77,7 +77,7 @@ pub fn jarník<M: Maze>(mut maze: M, rng: &mut Arengee) -> M {
             .filter(|n| !visited[maze.get_index(*n)])
             .collect::<Vec<_>>();
         if !possible_targets.is_empty() {
-            let target = possible_targets[rng.usize(0..possible_targets.len())];
+            let target = possible_targets[rng.u32(0..possible_targets.len() as u32) as usize];
             maze.carve(e, target);
             visited[maze.get_index(target)] = true;
             vertices.push(target);
